requirements.txt
django>=5.0.10,<6.0
djangorestframework>=3.16.0
mypy>=1.8.0
django-stubs>=4.2.7
djangorestframework-stubs>=3.15.2
python-decouple>=3.8

